Fixed Install-Frontend for Post
Long time ago we added functionality to delete the assets folder before installing the frontend. With Jura we changed the concept that we don't
delete the folder, but delete only the files which are in the NuGet package. Since Jura has '/images', '/js' and '/css' instead of '/assets' we forget to remove this lines.