added support for build-id in metadata.json
the problem is that since some time we have the problem that the padded
build number for Frontend nuget packags of release branches is not
correct.
The problem is that the build-version now contains stuff like 3.7
instead of the real build number so we add a new property which must
contain the build id.